+++ packages/tv-scraper.nix
@@ -1,18 +1,14 @@
 # ddotfiles/packages/tv.nix ⮞ https://github.com/QuackHack-McBlindy/dotfiles
-{ 
+{
   self,
   stdenv,
   lib,
   python3,
 } : let # 🦆 says ⮞ python dependencies
   pythonEnv = python3.withPackages (ps: [
-    ps.sounddevice
+    ps.lxml
     ps.requests
-    ps.python-dotenv
   ]);
-  cargoSrc = ./zigduck;
-  cargoToml = builtins.fromTOML (builtins.readFile (cargoSrc + "/Cargo.toml"));
-  cargoVersion = cargoToml.package.version;
 in  # 🦆 says ⮞ code source
 stdenv.mkDerivation {
     name = "tv-scraper";
@@ -32,7 +28,5 @@ stdenv.mkDerivation {
       description = "TV-scraper";
       license = lib.licenses.mit;
       maintainers = [ "QuackHack-McBlindy" ];
-      version = cargoVersion;
-      
-    };}
 
+    };}