+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Removed
+
+
+## [0.2.4] - 2026-09-14
+
+### Added
+
 - Added CI workflow.
 - `tv-scraper` python package that scrapes Nix defined TV-channel URLs.
 - The `zigduck-cli` package now supports `--version`.
@@ -22,7 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed the critical issue with Rust code missing.
 
 ### Removed
-
 
 ## [0.2.3] - 10-10-2026